Changed token loading and storing functions
Multiple entries can now be stored in the login json file.
The are stored under the account_type ('github') and username.
If a token already exists under the account_type/username the
token is simply overwritten.

A genterator has also been added to generate all the tokens of
a given account type.